Split the types of game objects: Either a simple object or a standard object
A simple object is one that is not rendered but may encompass some sort of game
processing logic.

A standard object is one that may be rendered but also has a location. The user
is required to specify a transform query function that returns the object's location.
From this we are able to move the shader variable updates from the client app into
the game engine.