Remove dependency on deprecated PGM library.
While we're at it, I also changed the depth map outputter to do the autoleveling
of the intensity values. This way we don't have to convert it every time we want
to see what's going on. I think this is an improvement, but obviously we still need
to figure out how to only produce this output on request and not every time we run
the examples...