New game state idea
Split the game state from the wires. Each frame, a game state is generated from the wires
and this state is then passed as a reader monad during the next frame. We also changed
game objects to be a transform along with a set of components. The currently available
components are a render object and a bounding volume. Eventually we would like the components
to be user-defined, but this probably isn't too much of a problem.

Also, there's a bit too much logic in UI/Lambency.hs... this should probably change.