Apply dithering to spotlight.frag
This was used in order to avoid banding artifacts on the plane below the
spotlight. The algorithm is fairly simple: once you have the final color
computed as a float, stochastically choose the nearest quantized value
based on how far away from it you are. In general this seems to produce
pretty good results for 8-bit framebuffers.