Fixed bug due to my misunderstanding of the difference between CURSOR_HIDDEN and CURSOR_DISABLED
That being said, this doesn't mean that glfw/glfw#153 is not still an issue. However,
even if CURSOR_HIDDEN is selected as the input mode, I would expect setCursorPos() followed
by getCursorPos() to correctly return the delta in mouse movement between the two calls. What
is happening is that only the y-component is updated. Knowing the difference between CURSOR_HIDDEN
and CURSOR_DISABLED, I don't think that this will ever be a problem in practice, but I assume
that the bug still exists...