Make sure that the cursor input is never too wacky.
This is another small hack to make it so that you can always find the scene once you've
started the application. Lambency just assumes that you want a hidden cursor and doesn't
really expose this to the user. There's probably a way to do this within FRP that creates
a wire that sets the proper behavior of a cursor and then constantly returns values that
adhere to that behavior.