Get rid of writer MonadT
It turns out that if you run an infinite evaluation of a monad writer
then you are `mappend`ing over and over again. Since you never evaluate
the result of the writer, or evaluate it when you quit, thunks are built
up until you run out of memory and the program crashes.