Small hack for MTL loader
If the MTL file contains zero'd out ambient reflectivity, but then also
gives a texture for ambient reflectivity, then assume that the texture is
the same as the diffuse texture, and set the reflectivity to V3 1 1 1. This
is a fairly small hack, as this should rarely be the case, although I'm worried
that in the future it will cause more problems than it solves.