Check faces to see what we're loading in the obj file.
Some OBJ files seem to allow mismatches with what objects in the file
have what defined. In other words, the same file could contain objects
that have texture coordinates and that don't have texture coordinates.
In this case, prior to outputting a mesh when parsing an OBJ file, set the
corresponding OBJ info based on what the data in the faces contains.