Bracket resources for the cube in the cube demo.
This makes me notice that relying on inhibition to unload resources is not
enough. When a wire inhibits, there is no guarantee that downstream wires will
run at all, meaning they could never trigger the proper resource unloading.

The likely solution to this is to store a tagged list of IO ()'s that correspond
to the unloading routine for resources. Each frame, we "touch" the resources
that we still need, meaning at the end of the frame we can run all of the IO
programs that didn't get touched. Unclear what the performance implications of
this would be. Hopefully not too bad...