Fix additional memory leak.
In particular, the game object being fed back into the game loop need to be
evaluated every frame. In some scenarios (such as the demos) we simply pass the
unit value `()` as the game object. Since this value is never needed, it's never
evaluated, and hence we just keep building up thunks to evaluate it.