Add initial version of required packets protocol.
This doesn't actually hook up the user API for specifying whether or not a
packet is required, but it does actually provide logic for dropping packets
that arrive out of order. The demo example is *way* smoother now, which is
kinda cool.

I'm not convinced that this logic is bug-free, though, and without better
testing it wouldn't be a good idea to keep it in the state its in.