Add packet queues.
Also changed the semantics of ACK-ing packets. I think this is still
wrong though. On the server we're acking per-player. This is good, but
we should be avoiding doing so as a matter of choosing which packets
are good or not. Ideally we should ack ALL packets and then have a queue
of packets that we apply regardless of the higher-level semantics (such
as packet queues). I worry that this gets hairy fast, though, so I'm not
totally sure how we're supposed to handle this.